requirements.txt
decorator==3.4.0
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pbr==0.10.8
six==1.9.0
SQLAlchemy==0.9.8
sqlalchemy-migrate==0.9.5
sqlparse==0.1.14
Tempita==0.5.2
Werkzeug==0.10.1
